// frontend-admin/types/user.d.ts
export type User = { 
	id: string; 
	email?: string; 
	name?: string; 
	password?: string; 
	phone?: string; 
	bio?: string; 
	avatar?: string; 
	cover?: string; 
	date_of_birth?: string; 
	sex?: string; 
	goal?: string; 
	height_cm?: number; 
	weight_kg?: number; 
	status?: string;
};

export type UserPayload = { 
	email?: string; 
	name?: string; 
	password?: string; 
	phone?: string; 
	bio?: string; 
	avatar?: string; 
	cover?: string; 
	date_of_birth?: string; 
	sex?: string; 
	goal?: string; 
	height_cm?: number; 
	weight_kg?: number; 
	status?: string;
};